requirements.txt
tensorflow
keras
os
re
scipy
skimage.transform
matplotlib
numpy
matplotlib.pyplot
